requirements.txt
torch==1.8
torchvision==0.9.0
numpy
pillow
tensorboardX
opencv-python
yacs
tqdm
scipy
matplotlib
progressbar2
imageio
eniops
plyfile